+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Fix `bit-and` in internal-sequencer
 - Support `##Inf` and `##-Inf` conversion to float within ugens
 - Remove `INFINITY` var
+- Fix race conditions in `control-proxy-value-atom`